Button refactor to JSX, moved logic computed code
