+++ src/components/GBtn/__story__/1-Button.stories.js
@@ -1,6 +1,7 @@
 import { text, withKnobs, boolean, number, array } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import GBtn from '../GBtn';
+import GBtnNew from '../GBtnNew';
 
 export default {
   title: 'Button',
@@ -17,7 +18,7 @@ export const test1 = () => ({
 })
 
 export const GBtnDefault = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     text: { default: text('Button Text', 'NORMAL BUTTON') },
     flat: { type: Boolean, default: boolean('flat', false) },
@@ -36,11 +37,11 @@ export const GBtnDefault = () => ({
     right: { type: Boolean, default: boolean('right', false) },
     tile: { type: Boolean, default: boolean('tile', false) },
   },
-  template: `<g-btn :tile="tile" :flat="flat" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled" :block="block" :elevation="elevation" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{text}}</g-btn>`,
+  template: `<g-btn-new :tile="tile" :flat="flat" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled" :block="block" :elevation="elevation" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{text}}</g-btn-new>`,
 })
 
 export const GBtnOutlined = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     text: { default: text('Button Text', 'OUTLINED BUTTON') },
     xSmall: { type: Boolean, default: boolean('xSmall', false) },
@@ -60,11 +61,11 @@ export const GBtnOutlined = () => ({
     tile: { type: Boolean, default: boolean('tile', false) },
 
   },
-  template: `<g-btn :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled"  :block="block" :text-color="textColor" :outlined="outlined" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{text}}</g-btn>`,
+  template: `<g-btn-new :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled"  :block="block" :text-color="textColor" :outlined="outlined" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{text}}</g-btn-new>`,
 })
 
 export const GBtnDepressed = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     text: { default: text('Button Text', 'DEPRESSED BUTTON') },
     xSmall: { type: Boolean, default: boolean('xSmall', false) },
@@ -84,11 +85,11 @@ export const GBtnDepressed = () => ({
     tile: { type: Boolean, default: boolean('tile', false) },
 
   },
-  template: `<g-btn :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled"  :block="block" :text-color="textColor" :depressed="depressed" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{text}}</g-btn>`,
+  template: `<g-btn-new :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled"  :block="block" :text-color="textColor" :depressed="depressed" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{text}}</g-btn-new>`,
 })
 
 export const GBtnText = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     textValue: { default: text('Button Text', 'TEXT BUTTON') },
     xSmall: { type: Boolean, default: boolean('xSmall', false) },
@@ -108,12 +109,12 @@ export const GBtnText = () => ({
     tile: { type: Boolean, default: boolean('tile', false) },
 
   },
-  template: `<g-btn :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled" :block="block" :text="text" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{textValue}}</g-btn>`,
+  template: `<g-btn-new :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled" :block="block" :text="text" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{textValue}}</g-btn-new>`,
 })
 
 
 export const GBtnRounded = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     textValue: { default: text('Button Text', 'ROUNDED BUTTON') },
     xSmall: { type: Boolean, default: boolean('xSmall', false) },
@@ -134,12 +135,12 @@ export const GBtnRounded = () => ({
     tile: { type: Boolean, default: boolean('tile', false) },
 
   },
-  template: `<g-btn :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled"  :block="block" :background-color="backgroundColor" :rounded="rounded" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{textValue}}</g-btn>`,
+  template: `<g-btn-new :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled"  :block="block" :background-color="backgroundColor" :rounded="rounded" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">{{textValue}}</g-btn-new>`,
 })
 
 
 export const GBtnFab = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     textValue: { default: text('Button Text', 'ROUNDED BUTTON') },
     xSmall: { type: Boolean, default: boolean('xSmall', false) },
@@ -159,13 +160,13 @@ export const GBtnFab = () => ({
     tile: { type: Boolean, default: boolean('tile', false) },
 
   },
-  template: `<g-btn :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled" :fab="fab" :background-color="backgroundColor" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">
+  template: `<g-btn-new :tile="tile" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :disabled="disabled" :fab="fab" :background-color="backgroundColor" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">
               <i class="material-icons g-icon">android</i>
-            </g-btn>`,
+            </g-btn-new>`,
 })
 
 export const GBtnIcon = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     textValue: { default: text('Button Text', 'ROUNDED BUTTON') },
     xSmall: { type: Boolean, default: boolean('xSmall', false) },
@@ -183,13 +184,13 @@ export const GBtnIcon = () => ({
     right: { type: Boolean, default: boolean('right', false) },
 
   },
-  template: `<g-btn :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :icon="icon" :disabled="disabled" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">
+  template: `<g-btn-new :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right" :icon="icon" :disabled="disabled" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">
               <i class="material-icons g-icon">star</i>
-            </g-btn>`,
+            </g-btn-new>`,
 })
 
 export const GBtnGradient = () => ({
-  components: { GBtn },
+  components: { GBtnNew },
   props: {
     textValue: { default: text('Button Text', 'GRADIENT BUTTON') },
     xSmall: { type: Boolean, default: boolean('xSmall', false) },
@@ -208,9 +209,9 @@ export const GBtnGradient = () => ({
     gradientAngle: { type: String, default: text('gradientAngle', '120deg') },
     tile: { type: Boolean, default: boolean('tile', false) },
   },
-  template: `<g-btn :tile="tile" :gradient-angle="gradientAngle" :gradient="gradient" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right"  :disabled="disabled" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">
+  template: `<g-btn-new :tile="tile" :gradient-angle="gradientAngle" :gradient="gradient" :absolute="absolute" :fixed="fixed" :top="top" :bottom="bottom" :left="left" :right="right"  :disabled="disabled" :text-color="textColor" :x-small="xSmall" :small="small" :large="large" :x-large="xLarge">
             {{textValue}}
-            </g-btn>`,
+            </g-btn-new>`,
 })
 
 export const test2 = () => ({
